requirements.txt
absl-py==2.3.1
accelerate==1.9.0
adapters==1.2.0
adjustText==1.3.0
affine==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
albumentationsx==2.0.9
annotated-types==0.7.0
anyio==4.11.0
asciitree==0.3.3
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
augmentations==0.0.1
autocorrect==2.6.1
beautifulsoup4==4.13.4
blake3==1.0.8
blinker==1.9.0
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.1
comm==0.2.2
compressed-tensors==0.11.0
conch @ git+https://github.com/MahmoodLab/CONCH.git@141cc09c7d4ff33d8eda562bd75169b457f71a62
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
Cython==3.1.6
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
entrypoints==0.3
et_xmlfile==2.0.0
executing==2.2.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
ffmpy==0.6.1
filelock==3.20.0
flake8==3.7.9
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2024.6.1
ftfy==6.3.1
gdown==5.2.1
geojson==3.1.0
geopandas==1.0.1
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1641732907644/work
grpcio==1.74.0
h5py==3.15.1
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
imagecodecs==2025.3.30
imageio==2.37.0
importlib_resources==6.5.2
inexactsearch==1.0.2
inquirerpy==0.3.4
interegular==0.3.3
isort==4.3.21
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
lightning-utilities==0.15.2
littleutils==0.2.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
Markdown==3.8.2
MarkupSafe==3.0.3
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.6.1
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.13.0
numba==0.61.2
numcodecs==0.13.1
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
nvitop==1.5.3
openai==2.5.0
openai-harmony==0.0.4
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
openslide-python==1.4.3
orjson==3.11.3
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peft==0.17.0
pfzy==0.3.4
pillow==12.0.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.31.1
psutil==7.1.3
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.4.2
pycocotools==2.0.11
pycodestyle==2.5.0
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
pydeck==0.9.1
pydub==0.25.1
pyflakes==2.1.1
pynndescent==0.5.13
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.1
PySocks==1.7.1
pyspellchecker==0.8.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
ray==2.50.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rpds-py==0.27.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
-e git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf#egg=segment_anything
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.42.0
setproctitle==1.3.7
shapely==2.0.6
silpa_common==0.3
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundex==1.1.3
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
stack-data==0.6.3
starlette==0.48.0
stringzilla==4.0.14
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.21.0a20250801
tenacity==9.1.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
timm==1.0.19
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.12.0
torch==2.10.0
torchmetrics==1.8.2
torchvision==0.25.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.4.0
turfpy==0.0.8
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
umpy==0.1.0
universeg @ git+https://github.com/JJGO/UniverSeg.git@833a0c34c65e38d675e21bd48ddec6797cc03259
UNKNOWN @ git+https://github.com/Mahmoodlab/CONCH.git@141cc09c7d4ff33d8eda562bd75169b457f71a62
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
xgrammar==0.1.25
yacs==0.1.8
yapf==0.29.0
yarl==1.22.0
zarr==2.18.3
